Build an Ad Library Query

This will build a list that can be passed to adlib_get. Visit https://developers.facebook.com/docs/marketing-api/reference/ads_archive/ for full API documentation.

adlib_build_query(
  ad_reached_countries,
  ad_active_status = c("ALL", "ACTIVE", "INACTIVE"),
  ad_delivery_date_max = NULL,
  ad_delivery_date_min = NULL,
  ad_type = c("POLITICAL_AND_ISSUE_ADS", "HOUSING_ADS", "NEWS_ADS",
    "UNCATEGORIZED_ADS", "ALL"),
  bylines = NULL,
  delivery_by_region = NULL,
  potential_reach_max = NULL,
  potential_reach_min = NULL,
  publisher_platform = "FACEBOOK",
  search_page_ids = NULL,
  search_terms = NULL,
  limit = 1000,
  fields = "ad_data"
)

Arguments

ad_reached_countries
Vector of ISO country codes. Facebook delivered the ads in these countries.

ad_active_status
Status of the ads. One of 'ACTIVE', 'INACTIVE', or 'ALL'

ad_delivery_date_max
character. Search for ads delivered before the date (inclusive) you provide. The date format should be YYYY-mm-dd.

ad_delivery_date_min
character. Search for ads delivered after the date (inclusive) you provide. The date format should be YYYY-mm-dd.

ad_type
The type of ad. One of 'ALL', 'POLITICAL_AND_ISSUE_ADS', 'HOUSING_ADS', 'NEWS_ADS', or 'UNCATEGORIZED_ADS'. Currently only 'POLITICAL_AND_ISSUE_ADS' is supported.

bylines
Filter results for ads with a paid for by disclaimer byline, such as political ads that reference “immigration” paid for by “ACLU”.

delivery_by_region
Character vector of region names. View ads by the region (such as state or province) where people live or were located when they saw them.

potential_reach_max
Search for ads with a maximum potential reach. Must be one of these range boundaries: 1000, 5000, 10000, 50000, 100000, 500000, 1000000, or leave empty for no maximum boundary.

potential_reach_min
Search for ads with a minimum intended reach. Must be one of these range boundaries: 100, 1000, 5000, 10000, 50000, 100000, 500000, 1000000.

publisher_platform
The platform on which the ads appeared. One or more of "FACEBOOK", "INSTAGRAM", "AUDIENCE_NETWORK", "MESSENGER", "WHATSAPP".

search_page_ids
A vector of up to 10 page IDs to search.

search_terms
A search string.

limit
The maximum number of results to return

fields
the fields to include in the response. See details for values.

Value

A list of params

Details

Preset groups of fields can be specified by "ad_data", "demographic_data", or "region_data". Otherwise, you can pick and choose fields from the following list. The only *required* field when picking and choosing is ad_snapshot_url, since that's the unique identifier for each ad.

ad_data - Choose this to include all columns for an ad_table table

demographic_data - Choose this to include all columns for demographic_table

region_data - Choose this to include all columns for a region_table

ad_creation_time

ad_creative_body

ad_creative_link_caption

ad_creative_link_description

ad_creative_link_title

ad_delivery_start_time

ad_delivery_stop_time

ad_snapshot_url

currency

demographic_distribution

funding_entity

impressions

page_id

page_name

potential_reach

publisher_platforms

region_distribution

spend