Build an Ad Library Query

This will build a list that can be passed to adlib_get. Visit https://developers.facebook.com/docs/marketing-api/reference/ads_archive/ for full API documentation.

adlib_build_query(
  ad_reached_countries,
  ad_active_status = c("ACTIVE", "INACTIVE", "ALL"),
  ad_type = c("POLITICAL_AND_ISSUE_ADS", "HOUSING_ADS", "NEWS_ADS", "UNCATEGORIZED_ADS",
    "ALL"),
  bylines = NULL,
  impression_condition = c("HAS_IMPRESSIONS_LIFETIME", "HAS_IMPRESSIONS_YESTERDAY",
    "HAS_IMPRESSIONS_LAST_7_DAYS", "HAS_IMPRESSIONS_LAST_30_DAYS",
    "HAS_IMPRESSIONS_LAST_90_DAYS"),
  publisher_platform = "FACEBOOK",
  search_page_ids = NULL,
  search_terms = NULL,
  limit = 1000,
  fields = "ad_data"
)

Arguments

ad_reached_countries
Vector of ISO country codes. Facebook delivered the ads in these countries.

ad_active_status
Status of the ads. One of 'ACTIVE', 'INACTIVE', or 'ALL'.

ad_type
The type of ad. One of 'ALL', 'POLITICAL_AND_ISSUE_ADS', 'HOUSING_ADS', 'NEWS_ADS', or 'UNCATEGORIZED_ADS'. Currently only 'POLITICAL_AND_ISSUE_ADS' is supported.

bylines
Filter results for ads with a paid for by disclaimer byline, such as political ads that reference “immigration” paid for by “ACLU”.

impression_condition
When the ad most recently had impressions. One of "HAS_IMPRESSIONS_LIFETIME", "HAS_IMPRESSIONS_YESTERDAY", "HAS_IMPRESSIONS_LAST_7_DAYS", "HAS_IMPRESSIONS_LAST_30_DAYS", "HAS_IMPRESSIONS_LAST_90_DAYS"

publisher_platform
The platform on which the ads appeared. One or more of "FACEBOOK", "INSTAGRAM", "AUDIENCE_NETWORK", "MESSENGER", "WHATSAPP".

search_page_ids
A vector of up to 10 page IDs to search.

search_terms
A search string.

limit
The maximum number of results to return

fields
the fields to include in the response. See details for values.

Value

A list of params

Details

Fields can be one of the following values.

ad_data - Choose this to include all columns for an ad_table table

demographic_data - Choose this to include all columns for demographic_table

region_data - Choose this to include all columns for a region_table

ad_creation_time

ad_creative_body

ad_creative_link_caption

ad_creative_link_description

ad_creative_link_title

ad_delivery_start_time

ad_delivery_stop_time

ad_snapshot_url

currency

demographic_distribution

funding_entity

impressions

page_id

page_name

publisher_platforms

region_distribution

spend